Exercícios + testes

+++ exe05/main.c
@@ -6,17 +6,17 @@ Saida: Informa se o aluno foi aprovado caso sua media seja maior ou igual a 6.
 */
 int main(){
 
-    float notas[4];
+    int notas[4];
     float soma=0;
     float media;
 
     for(int i=0; i<4; i++){
         // printf("Informe sua nota: ");
-        scanf("%f",&notas[i]);
+        scanf("%d",&notas[i]);
         soma += notas[i];
     }
 
-    media = soma/4;
+    media = soma/4.0;
 
     if(media>=6){
         printf("NOTA = %.1f (APROVADO)",media);
